refactor: Fix tests
